Convierte la constante EPSG a Long para ProjectionFactory.getProjection
Tercera ronda de CI: getProjection tiene sobrecargas (Long) y (String), y
ProjectionConstants.EPSG_WORLD_GEODETIC_SYSTEM es int de Java (Int en
Kotlin), que no se amplía implícitamente a Long. Único error restante de
compileDebugKotlin.

+++ mobile/android/app/src/main/java/ec/cnel/sgaie/movil/data/GeoPackageProvider.kt
@@ -67,7 +67,8 @@ object GeoPackageProvider {
     ) {
         if (geoPackage.isFeatureTable(nombreTabla)) return
 
-        val proyeccion = ProjectionFactory.getProjection(SRS_WGS84)
+        // getProjection recibe long; la constante EPSG es int en Java.
+        val proyeccion = ProjectionFactory.getProjection(SRS_WGS84.toLong())
         val srs: SpatialReferenceSystem = geoPackage.spatialReferenceSystemDao.getOrCreate(proyeccion)
 
         val geometryColumns = GeometryColumns()
